Refactor: KanbanBoard 컴포넌트 분리 #39
